refactor(tools): split experimental tool gates into focused modules
